+++ matrix.h
@@ -4,14 +4,14 @@
 typedef struct Matrix {
     int rows;
     int cols;
-    double* data;
+    float* data;
 } Matrix;
 
 Matrix* matrix_create(int rows, int cols);
 
 Matrix* matrix_from_shape(Matrix* m);
 
-Matrix* matrix_from_data(int rows, int cols, double* data);
+Matrix* matrix_from_data(int rows, int cols, float* data);
 
 void matrix_destroy(Matrix* m);
 
@@ -25,13 +25,13 @@ Matrix* matrix_sub(Matrix* a, Matrix* b);
 
 Matrix* matrix_mul(Matrix* a, Matrix* b);
 
-Matrix* matrix_divf(Matrix* a, double f);
+Matrix* matrix_divf(Matrix* a, float f);
 
-Matrix* matrix_subf(Matrix* m, double f);
+Matrix* matrix_subf(Matrix* m, float f);
 
-Matrix* matrix_mulf(Matrix* m, double f);
+Matrix* matrix_mulf(Matrix* m, float f);
 
-double matrix_sum(Matrix* m);
+float matrix_sum(Matrix* m);
 
 void matrix_print(char* label, Matrix* m, int y, int x);
 
@@ -41,7 +41,7 @@ Matrix* matrix_relu(Matrix* m);
 
 Matrix* matrix_softmax(Matrix* m);
 
-Matrix* matrix_one_hot(const double* Y, int len);
+Matrix* matrix_one_hot(const float* Y, int len);
 
 Matrix* matrix_cols(Matrix* m, int start, int end);
